Create a C submodule that has all wrapped functions as-is.
Add an __init__ function that always runs GDALAllRegister on startup.
Add tests that one ore more drivers exist.

Use GDAL's custom error handler funcitonality:
all errors internally reported to CPLError() are now handled by a julia function
currently this function always throws an error such that we don't need to check for them

For other functions such as GetDriverByName, when a driver cannot be found,
this is not reported to CPLError() and a NULL pointer is returned.
In this case we ofter still want an error, that is what checknull() does.