ensure GDAL_DATA is set
If it is not set as an environment variable already,
we try to find it ourselves by trying to use `gdal-config`
on Unix, and looking relative to the shared library path
otherwise. This is done in `__init__`, every time the
module loads.

If this fails, the user gets a warning on loading the module.
Tests will fail, but parts of GDAL.jl that don't need GDAL_DATA
can still be used.