relax null checking
Fixes #19

Now a GDALError is only thrown in `checknull` if both the pointer is null and the last error type is either CE_Failure or CE_Fatal, as suggested by Even Rouault.

The error handler that we register is given the same behavior (throw on failure & fatal), instead of previously always raising an error when called.

At the same time this restructures and tests the error handling a bit better.