1.0 Release with latest GDAL3 bindings (#73)
* break everything

1. stop removing prefixes from GDAL function names, thereby overloading them
2. removes the C submodule, everything is not together in the main module

I realize the function renaming is quite breaking. In the README.md I wrote
a bit about the changes and how to update code with the [renamer](https://gist.github.com/visr/0a2ad3fe92073345c93c2ca42f5f58a0#file-renamer-jl).

Before we had many methods of (for instance) destroy. We used multiple dispatch to select the right C function if you called GDAL.destroy(x),
depending on the type of x:

```json
"destroy": [
        "gdaldestroy",
        "gdal_cg_destroy",
        "ogr_fld_destroy",
        "ogr_gfld_destroy",
        "ogr_fd_destroy",
        "ogr_f_destroy",
        "ogr_ds_destroy",
        "ogr_sm_destroy",
        "ogr_st_destroy",
        "ogr_stbl_destroy"
    ]
```

This only worked because we rewrote the wrapping code, to pretend that `Ptr{Cvoid}` was actually `Ptr{OGRFeatureH}`, where `OGRFeatureH` was an abstract type
we put in https://github.com/JuliaGeo/GDAL.jl/blob/b75ac1c3a351ac934e35827289a0d522a1fa7005/src/types.jl. While it was nice to use dispatch here,
it did not match well with the C API. Firstly, it meant that wrapping became a complicated puzzle which did not neccesarily work. In `types.jl` we used
subtyping to simulate that some functions could take different types of inputs. Now it is up to the user to select the right function. The functions
will accept everything that can be converted to the types that C expects. This is in line with the new default Clang.jl behavior. This more straightforward
wrapping of the GDAL C API is a better fit for this package, allowing others such as ArchGDAL.jl to try to make it more Julian.

Regarding the C submodule, since we now do less rewriting, there is no strong reason anymore to keep them separate, folding them together.

* address review comments

* add link to renamer.jl

* EzXML fix deprecation warning

* Set release to 1.0.0